Move custom tags to its own file
